src/retriever/requirements.txt
flask==2.3.3
llama-index-core==0.12.22
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
llama-index-embeddings-huggingface==0.5.2
llama-index-vector-stores-postgres==0.4.2